third_party/blink/web_tests/platform/win11-arm64/webaudio/Analyser/realtimeanalyser-fftsize-reset-expected.txt

This is a testharness.js-based test.
PASS # AUDIT TASK RUNNER STARTED.
PASS Executing "128->1024"
PASS Executing "512->256"
PASS Audit report
PASS > [128->1024] 
PASS   Time data equals [0.6596450209617615,0.6688088178634644,0.7413902878761292,0.7453146576881409,0.737322986125946,0.9535815119743347,0.2407812476158142,-0.8548296689987183,-0.8268423676490784,-0.7138630151748657,-0.7638787031173706,-0.6886090636253357,-0.6578097939491272,-0.6554352641105652,-0.5911034941673279,-0.5760258436203003...] with an element-wise tolerance of {"absoluteThreshold":0,"relativeThreshold":0}.
FAIL X Linear FFT data after setting fftSize = 1024 does not equal [6.063636419639806e-7,6.10060567396431e-7,6.514114261335635e-7,6.956996116969094e-7,7.994843258529727e-7,8.804310027699103e-7,0.000001099945507121447,0.0000013056857142146328,0.000001665369950387685,0.000002156581331291818,0.000002902891083067516,0.000004041599822812714,0.000005879006039322121,0.000008955552402767353,0.000013970826330478303,0.000019471892301226035...] with an element-wise tolerance of {"absoluteThreshold":0,"relativeThreshold":0.0000019455}.
	Index	Actual			Expected		AbsError		RelError		Test threshold
	[489]	5.5076754268412742e-10	5.5076948557442051e-10	1.9428902930940239e-15	3.5275924755847765e-6	1.0715220341850350e-15
	Max AbsError of 1.9428902930940239e-15 at index of 489.
	Max RelError of 3.5275924755847765e-6 at index of 489.
 assert_true: expected true got false
FAIL < [128->1024] 1 out of 2 assertions were failed. assert_true: expected true got false
PASS > [512->256] 
PASS   Time data equals [-0.6523512601852417,-0.5852614641189575,-0.5780778527259827,-0.5480099320411682,-0.49195218086242676,-0.4970853626728058,-0.43464046716690063,-0.4184177815914154,-0.3903508484363556,-0.3461340367794037,-0.32595396041870117,-0.29549989104270935,-0.2502514719963074,-0.24035859107971191,-0.19149018824100494,-0.165742889046669...] with an element-wise tolerance of {"absoluteThreshold":0,"relativeThreshold":0}.
PASS   Linear FFT data after setting fftSize = 256 equals [0.000007047586223052349,0.000015348627130151726,0.00010122780076926574,0.004589741118252277,0.01776057854294777,0.021656673401594162,0.009140600450336933,0.0005019662203267217,0.004027072805911303,0.010545582510530949,0.009361358359456062,0.0027135838754475117,0.0004940525395795703,0.004096011631190777,0.007559461053460836,0.0048733921721577644...] with an element-wise tolerance of {"absoluteThreshold":0,"relativeThreshold":0.0000018592}.
PASS < [512->256] All assertions passed. (total 2 assertions)
FAIL # AUDIT TASK RUNNER FINISHED: 1 out of 2 tasks were failed. assert_true: expected true got false
Harness: the test ran to completion.
